delete duplicate login page
