+++ src/routes/MyAccount/MyAccount.css
@@ -1,0 +1,32 @@
+.account-menu {
+  display: flex;
+  justify-content: flex-end;
+  margin-right: 4rem;
+}
+
+.account-content {
+  display: flex;
+  justify-content: space-around;
+  margin: 4rem;
+  padding: 1rem;
+}
+
+.welcome-message {
+  margin: 0rem 1rem;
+  font-weight: 300;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+@media only screen and (max-width: 991px) {
+  .account-menu {
+    margin-right: 0;
+  }
+  .account-content {
+    flex-direction: column;
+    justify-content: center;
+    align-items: flex-start;
+    margin: 0;
+  }
+}
